requirements.txt
requests
pypdf
sentence_transformers
langchain_text_splitters
numpy
python-dotenv
huggingface-hub
streamlit